Move `AnimatedButton` higher
